requirements.txt
absl-py==0.11.0
adal==1.2.5
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1605217004767/work
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
azure-cognitiveservices-search-imagesearch==2.0.0
azure-common==1.1.26
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache==1.6.1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1600454382015/work
blis==0.4.1
brotlipy==0.7.0
cachetools==4.2.1
catalogue @ file:///home/conda/feedstock_root/build_artifacts/catalogue_1605613584677/work
certifi==2020.11.8
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1602537222527/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1602255302199/work
click==7.1.2
cloudpickle==1.6.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1604179079864/work
cycler==0.10.0
cymem @ file:///home/conda/feedstock_root/build_artifacts/cymem_1604919356432/work
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
detectron2==0.3+cu110
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
fastai @ file:///home/jhoward/anaconda3/conda-bld/fastai_1604932388885/work
fastai2==0.0.30
fastbook==0.0.14
fastcore @ file:///home/jhoward/anaconda3/conda-bld/fastcore_1605551797847/work
fastprocess==2.0.0
fastprogress @ file:///home/conda/feedstock_root/build_artifacts/fastprogress_1597932925331/work
filelock==3.0.12
future==0.18.2
fvcore==0.1.3.post20210218
gdown==3.12.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
graphviz==0.15
grpcio==1.35.0
h5py==3.1.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1602263269022/work
iopath==0.1.3
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1605455374814/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1604159561527/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1599554010055/work
isodate==0.6.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1605054524035/work
Jinja2==2.11.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1601671685479/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1602551949684/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1598486169312/work
jupyter-console @ file:///home/conda/feedstock_root/build_artifacts/jupyter_console_1598728807792/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1605735009305/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
Keras==2.4.3
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1604322295622/work
layoutlm @ file:///notebooks/unilm/layoutlm
lxml==4.5.1
Markdown==3.3.3
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1602267312178/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1605180228501/work
mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1605115651871/work
msrest==0.6.19
msrestazure==0.6.4
murmurhash==1.0.4
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1602859080374/work
nbconvert==5.6.1
nbdev==1.1.5
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1602732862338/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1605195931949/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1605103633466/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1604945996350/work
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.1.48
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pandas==1.1.4
pandocfilters==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1595548966091/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==7.1.2
plac==0.9.6
portalocker==2.2.1
preshed @ file:///home/conda/feedstock_root/build_artifacts/preshed_1605166129992/work
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1605543085815/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1605053337398/work
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydot==1.4.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1603558917696/work
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent @ file:///home/conda/feedstock_root/build_artifacts/pyrsistent_1605115595652/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1602326928339/work
pytesseract==0.3.7
python-dateutil==2.8.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1604321279890/work
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole @ file:///home/conda/feedstock_root/build_artifacts/qtconsole_1599147533948/work
QtPy==1.9.0
regex==2020.11.13
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1605186911681/work
requests-oauthlib==1.3.0
rsa==4.7.1
sacremoses==0.0.43
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1604232448678/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1604304779838/work
Send2Trash==1.5.0
sentencepiece==0.1.94
seqeval==0.0.12
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
spacy @ file:///home/conda/feedstock_root/build_artifacts/spacy_1605692171608/work
srsly @ file:///home/conda/feedstock_root/build_artifacts/srsly_1605085673973/work
tabulate==0.8.8
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.0
termcolor==1.1.0
terminado @ file:///home/conda/feedstock_root/build_artifacts/terminado_1602679586280/work
testpath==0.4.4
thinc @ file:///home/conda/feedstock_root/build_artifacts/thinc_1605620876750/work
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
tokenizers==0.10.1
torch==1.7.1
torchvision==0.8.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1604105045397/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1605543106900/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
transformers @ file:///notebooks/transformers
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1603125704209/work
utils==1.0.1
wasabi @ file:///home/conda/feedstock_root/build_artifacts/wasabi_1600272362626/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1605475534911/work
xmltodict==0.12.0
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1603668650351/work
